Separates availability form, logic and functions for updating to another component. HostProfile component becomes even smaller.
